feat: Remove HeatmapTempGraph component and clean up related imports in AnalyticsView
