+++ commitment/imports/ui/components/metrics-page/AnalyticsView.tsx
@@ -11,7 +11,6 @@ import { LeaderboardGraph } from "./LeaderboardGraph";
 // import { ContributionPieChart } from "./PieChartGraph";
 // import GraphCard from "./GraphCard";
 import HeatmapGraph from "./HeatMapGraph";
-import HeatmapTempGraph from "./HeatMapTempGraph";
 
 import { AnalyticsData } from "/imports/api/types";
 
@@ -188,10 +187,10 @@ export function AnalyticsView(): React.JSX.Element {
               />
             </div>
             <HeatmapGraph
-              data={analytics.metrics.contributors.heatMap} // Replace with real data
+              data={analytics.metrics.contributors.heatMap}
               title="Contributor Activity Heatmap"
             />
-            <HeatmapTempGraph data={analytics.metrics.contributors.heatMap}/>
+            
           </div>
         </div>
       </div>
